// src/lookups/enums/lookup-category.enum.ts
export enum LookupCategory {
  BLOOD_TYPE = 'BLOOD_TYPE',
  LIFESTYLE_HABITS = 'LIFESTYLE_HABITS',
  CHRONIC_CONDITIONS = 'CHRONIC_CONDITIONS',
  CHRONIC_CONDITION_CATEGORY = 'CHRONIC_CONDITION_CATEGORY',
  CHRONIC_CONDITION = 'CHRONIC_CONDITION',
  DISABILITY_TYPES = 'DISABILITY_TYPES',
  ALLERGY = 'ALLERGY',
  COMMON_SURGERIES = 'COMMON_SURGERIES',
  MEDICAL_SPECIALTY = 'MEDICAL_SPECIALTY',
  MEDICAL_SUB_SPECIALTY = 'MEDICAL_SUB_SPECIALTY',
}